# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NGOConnect AI resource coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Location"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated resource listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePage"
    post:
      operationId: createResource
      tags: [resources]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceInput"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getResource
      tags: [resources]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    patch:
      operationId: updateResource
      tags: [resources]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    delete:
      operationId: deleteResource
      tags: [resources]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted resource
...
components:
  schemas:
...
    Resource:
      type: object
      required: [id, ownerNgoId, ownerNgoName, name, category, quantity, unit, location, status, createdAt]
      properties:
        id: { type: integer }
        ownerNgoId: { type: integer }
        ownerNgoName: { type: string }
        name: { type: string }
        category:
          type: string
          enum: [food, medicines, healthcare_supplies, clothing, other]
        description: { type: string }
        quantity: { type: number }
        unit: { type: string }
        location: { type: string }
        expiryDate: { type: ["string", "null"], format: date }
        status:
          type: string
          enum: [available, reserved, depleted, expired]
        createdAt: { type: string, format: date-time }
    ResourceInput:
      type: object
      required: [name, category, quantity, unit, location]
      properties:
        name: { type: string, minLength: 2 }
        category:
          type: string
          enum: [food, medicines, healthcare_supplies, clothing, other]
        description: { type: string }
        quantity: { type: number, exclusiveMinimum: 0 }
        unit: { type: string, minLength: 1 }
        location: { type: string, minLength: 2 }
        expiryDate: { type: ["string", "null"], format: date }
    ResourceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        category:
          type: string
          enum: [food, medicines, healthcare_supplies, clothing, other]
        description: { type: string }
        quantity: { type: number, exclusiveMinimum: 0 }
        unit: { type: string, minLength: 1 }
        location: { type: string, minLength: 2 }
        expiryDate: { type: ["string", "null"], format: date }
        status:
          type: string
          enum: [available, reserved, depleted, expired]
...
    ResourcePage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
    Search:
      name: search
      in: query
      schema: { type: string }
    Category:
      name: category
      in: query
      schema:
        type: string
        enum: [food, medicines, healthcare_supplies, clothing, other]
    Location:
      name: location
      in: query
      schema: { type: string }
    Status:
      name: status
      in: query
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT