# shared/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval ðŸº
 * Do not edit manually.
 * Api
 * NGOConnect AI resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from './resourceCategory';
import type { ResourceStatus } from './resourceStatus';

export interface Resource {
  id: number;
  ownerNgoId: number;
  ownerNgoName: string;
  name: string;
  category: ResourceCategory;
  description?: string;
  quantity: number;
  unit: string;
  location: string;
  /** @nullable */
  expiryDate?: Date | null;
  status: ResourceStatus;
  createdAt: Date;
}