# shared/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval ðŸº
 * Do not edit manually.
 * Api
 * NGOConnect AI resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceInputCategory } from './resourceInputCategory';

export interface ResourceInput {
  /** @minLength 2 */
  name: string;
  category: ResourceInputCategory;
  description?: string;
  /** @exclusiveMinimum 0 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minLength 2 */
  location: string;
  /** @nullable */
  expiryDate?: Date | null;
}