# shared/api-zod/src/generated/types/resourcePage.ts
/**
 * Generated by orval ðŸº
 * Do not edit manually.
 * Api
 * NGOConnect AI resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from './resource';

export interface ResourcePage {
  items: Resource[];
  page: number;
  pageSize: number;
  total: number;
}