# shared/api-zod/src/generated/types/resourceUpdate.ts
/**
 * Generated by orval ðŸº
 * Do not edit manually.
 * Api
 * NGOConnect AI resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceUpdateCategory } from './resourceUpdateCategory';
import type { ResourceUpdateStatus } from './resourceUpdateStatus';

export interface ResourceUpdate {
  /** @minLength 2 */
  name?: string;
  category?: ResourceUpdateCategory;
  description?: string;
  /** @exclusiveMinimum 0 */
  quantity?: number;
  /** @minLength 1 */
  unit?: string;
  /** @minLength 2 */
  location?: string;
  /** @nullable */
  expiryDate?: Date | null;
  status?: ResourceUpdateStatus;
}